test: filter matched route
