+++ packages/runtime/src/runClientApp.tsx
@@ -5,8 +5,8 @@ import { createSearchParams } from 'react-router-dom';
 import Runtime from './runtime.js';
 import App from './App.js';
 import { AppContextProvider } from './AppContext.js';
-import type { AppContext, AppConfig, RouteItem, AppRouterProps, PageWrapper, RuntimeModules, InitialContext, RouteMatch } from './types';
-import { loadRouteModules, loadPageData, getPageConfig, matchRoutes } from './routes.js';
+import type { AppContext, AppConfig, RouteItem, AppRouterProps, PageWrapper, RuntimeModules, InitialContext } from './types';
+import { loadRouteModules, loadPageData, getPageConfig, matchRoutes, filterMatchesToLoad } from './routes.js';
 import { loadStyleLinks, loadScripts } from './assets.js';
 import { getLinks, getScripts } from './pageConfig.js';
 
@@ -189,31 +189,6 @@ async function loadNextPage(matches, preState) {
     pageData,
     pageConfig,
   };
-}
-
-function filterMatchesToLoad(matches, newMatches) {
-  let isNew = (match: RouteMatch, index: number) => {
-    // [a] -> [a, b]
-    if (!matches[index]) return true;
-
-    // [a, b] -> [a, c]
-    return match.route.id !== matches[index].route.id;
-  };
-
-  let matchPathChanged = (match: RouteMatch, index: number) => {
-    return (
-      // param change, /users/123 -> /users/456
-      matches[index].pathname !== match.pathname ||
-      // splat param changed, which is not present in match.path
-      // e.g. /files/images/avatar.jpg -> files/finances.xls
-      (matches[index].route.path?.endsWith('*') &&
-      matches[index].params['*'] !== match.params['*'])
-    );
-  };
-
-  return newMatches.filter((match, index) => {
-    return isNew(match, index) || matchPathChanged(match, index);
-  });
 }
 
 function getInitialContext() {
